[TR] TXM-3192 Capture the request path for play-auditing
Adds the request path to HeaderCarrier so that play auditing can
include it in audits without requiring developers to explicitly
pass through the request path to the audit call and without
adding play as a play-auditing dependency.

Despite the name, play-auditing no longer depends on play.
It will be renamed to auditing as part of the play-2.6 upgrade.

The path is added to 'otherHeaders' rather than a new field in
HeaderCarrier to avoid changing HeaderCarrier. HeaderCarrier is
in the http-core library which now depends on play-2.6 so a
change to HeaderCarrier would require consumers to upgrade to
play-2.6